replace uint64 and move to source and headers(old)

+++ include/bin2freq.h
@@ -6,11 +6,11 @@ class Bin2freq
 {
 public:
     static float calculate(float bin);
-    static void setWinSampRate(long int windowSize, double sampleRate);
+    static void setWinSampRate(long int windowSize, long int sampleRate);
     static unsigned long int getWindowSize();
     static unsigned long int getSampleRate();
 private:
     static unsigned long int windowSize;
-    static double sampleRate;
+    static unsigned long int sampleRate;
 };
 } // namespace tid
